// src/enums/category.ts
export enum Category {
  Electronics = 'Electronics',
  Appartment = 'Appartment',
  Furniture = 'Furniture',
  Vehicles = 'Vehicles',
  Clothing = 'Clothing',
  Books = 'Books',
  SportsAndOutdoors = 'Sports & Outdoors',
  HomeAppliances = 'Home Appliances',
  ToysAndGames = 'Toys & Games',
  Pets = 'Pets',
  Services = 'Services',
  Jobs = 'Jobs',
  BeautyAndHealth = 'Beauty & Health',
  GardenAndTools = 'Garden & Tools',
  RealEstate = 'Real Estate',
  ArtAndCollectibles = 'Art & Collectibles',
  MusicalInstruments = 'Musical Instruments',
  OfficeSupplies = 'Office Supplies',
  BabyAndKids = 'Baby & Kids',
  AutomotivePartsAndAccessories = 'Automotive Parts & Accessories',
  FoodAndBeverages = 'Food & Beverages',
  HobbiesAndCrafts = 'Hobbies & Crafts',
}